Create background.js
This code uses the `chrome.tabs.onUpdated` event listener to detect when the user navigates to a new tab or refreshes the current tab. It checks if the tab URL starts with https://www.kijijiautos.ca to determine if the user navigated to Kijiji Auto. It then extracts the search query parameters from the URL using the `URLSearchParams` object and logs them to the console.