Update popup.js
In the `handleSaveResultsClick` function, after adding the current page of search results to the `savedResults` array, we can display a confirmation message to the user that their results have been saved. We can do this by using the alert() function to display a message box with the confirmation text. 

Now, when the user clicks the "Save Results" button, they will see a message box with the text "Results saved!".